Lab 04 tests

+++ grade.sh
@@ -4,13 +4,14 @@ RED='\033[0;31m'
 NC='\033[0m'
 
 SECTION_SKIP=false
-DUE_DATE_UNIX=$(date -d 2021-10-30 +%s)
-DUE_DATE=$(date -d 2021-10-30 +"%a %b %d %H:%R")
+INSTRUCTOR_TESTS=true
+DUE_DATE_UNIX=$(date -d 2021-11-11 +%s)
+DUE_DATE=$(date -d 2021-11-11 +"%a %b %d %H:%R")
 
 ASSIGNMENT=$1
 if [ -z "$2" ]
 then
-	RSCDIR=grader_files
+	RSCDIR=.config
 else
 	RSCDIR=$2
 fi
@@ -18,6 +19,11 @@ fi
 if [[ $* == *-s* ]]
 then
 	SECTION_SKIP=true	
+fi
+
+if [[ $* == *-i* ]]
+then
+	INSTRUCTOR_TESTS=false
 fi
 
 ROSTER=$RSCDIR/roster.json
@@ -44,29 +50,44 @@ do
 		echo
 		make clean
 
-		echo "Replacing Main.ml"
-		cp main.ml old_main.ml
-		rm main.ml
-		cp ../../../$RSCDIR/$ASSIGNMENT/new_main.ml main.ml
+		if [ $INSTRUCTOR_TESTS != false ]
+		then
+			echo "Replacing Main.ml"
+			cp main.ml old_main.ml
+			rm main.ml
+			cp ../../../$RSCDIR/$ASSIGNMENT/new_main.ml main.ml
 
-		echo "Appending Instructor Tests"
-		cp $ASSIGNMENT.ml old_$ASSIGNMENT.ml
-		cat ../../../$RSCDIR/$ASSIGNMENT/instr_tests.ml >> $ASSIGNMENT.ml
-		pri=$(make $ASSIGNMENT | tee /dev/tty | cat | grep -c "\--------------")
-		[[ $pri -eq 4 ]] && p="-0 pts" || p="-5 pts"
-		echo "Print satement works properly      $p"
+			echo "Appending Instructor Tests"
+			cp $ASSIGNMENT.ml old_$ASSIGNMENT.ml
+			cat ../../../$RSCDIR/$ASSIGNMENT/instr_tests.ml >> $ASSIGNMENT.ml
+			pri=$(make $ASSIGNMENT | tee /dev/tty | cat | grep -c "function fun(x)")
+			[[ $pri -eq 1 ]] && p="-0 pts" || p="-5 pts"
+			echo "Print satement works properly      $p"
 
-		echo "Restoring Original main.ml"
-		rm main.ml
-		cp old_main.ml main.ml
-		rm old_main.ml
+			echo "Restoring Original main.ml"
+			rm main.ml
+			cp old_main.ml main.ml
+			rm old_main.ml
 
-		echo "Restoring Original $ASSIGNMENT.ml"
-		rm $ASSIGNMENT.ml
-		cp old_$ASSIGNMENT.ml $ASSIGNMENT.ml
-		rm old_$ASSIGNMENT.ml
+			echo "Restoring Original $ASSIGNMENT.ml"
+			rm $ASSIGNMENT.ml
+			cp old_$ASSIGNMENT.ml $ASSIGNMENT.ml
+			rm old_$ASSIGNMENT.ml
+			nvim $ASSIGNMENT.ml
+		else
+			make $ASSIGNMENT
+			make fib
+			cat answers.md
+			cat answers.txt
+			cat answers.tex
+		fi
 	else
-		echo -e "${RED}No Submission${NC}"
+		if [[ -f team.txt ]]
+		then
+			cat team.txt
+		else
+			echo -e "${RED}No Submission${NC}"
+		fi
 	fi
 	
 	TURN_IN_DATE=$(git log -1 --date=unix --format=%cd)
